[m0][mx] Fix test_override_config_table failed in m0/mx
